fix a problem with the arguments passed to the application added a fileFilterExclude list, the FileFilter ignores any files that contains those strings

+++ src/main/java/us/vicentini/mediamanager/actions/CopyFileAction.java
@@ -21,8 +21,9 @@ public class CopyFileAction {
         if(!destinationPath.exists()) {
             Files.createDirectories(destinationPath.toPath());
         }
-        
-        Files.copy(fromFile.toPath(), (new File(destinationPath, fromFile.getName())).toPath());
+        if(!(new File(destinationPath, fromFile.getName())).exists()) {
+            Files.copy(fromFile.toPath(), (new File(destinationPath, fromFile.getName())).toPath());
+        }
     }
  
     @Override
